Test/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
load-dotenv==0.1.0
pandas==1.3.5
psycopg2==2.9.9
